feat: rename DedupedTrait
